pyproject.toml
[project]
name = "amazing-iot-device"
version = "0.1.0"
description = "IoT device agent with Flask web interface"
authors = [
    {name = "Tewang.Chiu", email = "tewang.chiu@gmail.com"},
]
dependencies = [
    "flask>=2.3.0",
    "flask-login>=0.6.2",
    "flask-wtf>=1.1.1",
    "werkzeug>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    "paho-mqtt>=2.1.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
start = "python3 run.py"